Blurrr: I made way too many changes for this single commit. I thought this would be smaller. These changes redesign the semantics to more properly match the intended design of Zeroconf.
More specifically, a proper Zeroconf implementation should have:
Register/Unregister
StartDiscovery/StopDiscovery
StartResolve/EndResolve.

The problem with the original implementation was it combined/conflated
Discovery and Resolving into a single API. This is a bad idea because
Resolving is a more expensive operation that must continually create
traffic on the network. Additionally, the original implementation was
fetching IP addresses and ports immediately when these things should be
done at the last minute since these values can change over time. And the
original implementation was requesting IP Addresses, ports, and other
information from all services on the network instead of just the items you
actually want to resolve (which is way more expensive in terms of
traffic).

So this implementation breaks up the original implementation into separate
Discovery and Resolve functions.

I originally thought it would require fewer changes and code could be
shared between the two operations, but they were subtly different enough
that it was easier to create two separate files and let them do their own
thing.

I've also changed the parameters passed back in the callbacks and the API
parameters to be closer to the other implementations so integrating with
BlurrrZeroConf (or any other compat wrapper) should be easier.

I've added TXT record reporting in the resolve callback.
TODO: Add ability to set TXT records.

I am using host-order for the port. (I think.)

I originally tried to batch calls to take advantage of the Stop Enumeration
event, because it implied there was a sweep of all devices concluding with
the Stop. Some of the other implementations have a flag to hint the user
that more data is coming so hold off updating any GUIs for performance.
But it turns out that there is a 15-30sec delay between the discovery and
the Stop, so it didn't work out. I left it #define'd in, but the code has
already become stale.

I also left the legacy API code in place so it was easier to compare while
testing, but that code is contributing to the clutter, so I may need to
remove it.

main.cpp has a lot of hacks.
I was trying to figure out if FullName and all the escaping I did was
useful with the res_query equivalent. It does not appear to be. I used hex
instead of octal. I hope that is okay.

I also needed to switch to wstring's to get the console output not
mutulate the right apostrophe coming from iTunes "Al's Music".
So there was a lot of fighting/fixing string conversions, but I think
things are correct now.

There is a Microsoft design issue that worries me. It appears that you
must create a socket to start advertising. If so, I also worry that the
socket will conflict with your real service. I currently tell it to use a
random port for the socket, which is different than the advertisement
port. However, we still have a hidden socket that is waiting for a
connect which is stupid.

TODO: Remove the res_query dependencies (had to link in something extra).
- Add TXT Record support for setting.
- Clean up public header.
- Remove legacy code. (or #if it out)