Trying to fix sometimes exception/crash on shutdown which happens in the original example.
I find it helps to build as Release and quickly run through the program.
However, I have seen this in Debug too. I've also seen it happen not
running through the program quickly.

This exception is deep inside the Microsoft frameworks. It looks like it
has something to do with device enumeration.
I tried a bunch of different things.
- try/catch blocks in shutdown to catch the exception (didn't work)
- spin-loops to allow init to finish before quitting (didn't help)
- removing the callback delegates before calling stop(). (seemed to fix it)

I need to go back and clean up the other things that didn't help. Though
the init thing might be a good idea, but I probably should use proper
mutexes.