Removed mIsIniting because I don't think it actually worked. Also added dnssd_unintialize()
I don't think the isIniting was doing anything because there is already a
check for mServiceWatch which does most of the same thing.

Additionally, if the async task isn't finished yet (tested with
Sleep(10000), when I delete the wrapper object, the main object doesn't
seem to ever invoke the destructor. My suspicion is that the lambda
capture is keeping the object referenced and alive since the task is not
done.

I also tried to use std::mutex, but I couldn't figure out how to use it
with the async tasks. I kept getting an abort/exception when calling
unlock from the task.

In practice, I don't think this race condition will be an issue.
I think the mServiceWatch captures most of the edge case if it does
happen.